# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gym API",
        "description": "API pour la gestion de salle de sport",
        "contact": {
            "email": "admin@gym-gestion.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/api/categories/new": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Ajout de catégorie",
                "operationId": "7af26b6954e5f403dbd3ef29fef5be0b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_name"
                                ],
                                "properties": {
                                    "category_name": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "Mathematics"
                                    },
                                    "active_flag": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Enregistrement effectué"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Récupérer une catégorie par ID",
                "operationId": "5e3737f5cfd6dbb5742c61175b5fb0d4",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/categories/status/{id}": {
            "patch": {
                "tags": [
                    "Categories"
                ],
                "summary": "Modifier le statut d'une catégorie",
                "operationId": "f84ea164feba8ed3c99a174c22847c71",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statut modifié avec succès"
                    },
                    "404": {
                        "description": "Catégorie non trouvée"
                    }
                }
            }
        },
        "/api/categories/update/{id}": {
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Mettre à jour les informations d'une catégorie",
                "operationId": "09cffd484214f3378b117fa9c04fbf81",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categorie",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category_name"
                                ],
                                "properties": {
                                    "category_name": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "Mathematics"
                                    },
                                    "active_flag": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mis à jour effectuée avec succès"
                    },
                    "400": {
                        "description": "Demande invalide"
                    }
                }
            }
        },
        "/api/categories/delete/{id}": {
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Suppression d'une catégorie",
                "operationId": "7bd407cda39111b480c664d40e04385b",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/api/course-type/new": {
            "post": {
                "tags": [
                    "Course Types"
                ],
                "summary": "Ajouter un nouveau type de cours",
                "operationId": "33d62760ddccce51f1486352c0478041",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_type_name"
                                ],
                                "properties": {
                                    "course_type_name": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "Mathematics"
                                    },
                                    "active_flag": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Création avec succès"
                    },
                    "400": {
                        "description": "Demande invalide"
                    }
                }
            }
        },
        "/api/course-type/status/{id}": {
            "patch": {
                "tags": [
                    "Course Types"
                ],
                "summary": "Modifier le statut d'un type de cours",
                "operationId": "d5ea87b5ea21c181281addf97edc3682",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/course-type/update/{id}": {
            "put": {
                "tags": [
                    "Course Types"
                ],
                "summary": "Mettre à jours un type de cours",
                "operationId": "eb56a1bc1ad49b2d1090a8bf31f8a329",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la salle",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "course_type_name"
                                ],
                                "properties": {
                                    "course_type_name": {
                                        "type": "string",
                                        "maxLength": 45,
                                        "example": "Mathematics"
                                    },
                                    "active_flag": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Mis à jour effectuée avec succès"
                    },
                    "400": {
                        "description": "Demande invalide"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}